\docType{class}
\name{activity-class}
\alias{activity-class}
\title{An S4 class to represent a running activity.}
\description{
  An S4 class to represent a running activity.
}
\section{Slots}{

\describe{
\item{\code{actnr}}{number of the activity}

\item{\code{time}}{time of the measure points}

\item{\code{latitude}}{latitude value of the time point}

\item{\code{longitude}}{longitude value of the time point}

\item{\code{altitude}}{altitude value of the time point}

\item{\code{altitude_range}}{altitude range of the activity}

\item{\code{total_climb}}{total climb of the activity}

\item{\code{total_descent}}{total descent of the activity}

\item{\code{distance}}{distance in m between two time points}

\item{\code{total_distance}}{total distance in km of the activity}

\item{\code{heart_rate}}{heart rate of the time point}

\item{\code{speed}}{speed in km/h of the time point}

\item{\code{cadence}}{cadence in steps per minute of the time point}

\item{\code{total_time}}{total time in seconds of the activity}

\item{\code{duration}}{time in seconds between two time points}
}}

