\name{actlist}
\alias{actlist}
\title{
  makes a list of objects from class activity
}
\description{
  Get a list of dataframes from rds or tcx file. It creates then a list with new activity object in it.
}
\usage{
  actlist(actls)
}
\arguments{
  \item{actls}{
    list of dataframes created from rds or tcx file
  }
}
\value{
  \item{ls}{
    list of new objects from class activity
  }
}
