\docType{methods}
\name{addact}
\alias{addact}
\alias{addact,training-method}
\title{training.}
\usage{
addact(train, act)

\S4method{addact}{training}(train, act)
}
\arguments{
  \item{train}{an object of class training}

  \item{act}{an object of class activity}
}
\value{
  \item{train}{
    An object of class training
  }
}
\description{
  Get an object of the class activity. Add this object to the one object of the class training
}
