\docType{methods}
\name{plot_performance-methods}
\alias{plot_performance-methods}
\alias{plot_performance-methods, training-method}
\title{Plot heart rate, speed and altitude over time}
\usage{
  trimp_exp(obj, num = 1)

\S4method{plot_starts}{training}(obj, num)
}
\arguments{
  \item{obj}{
    an object of class training
  }
  \item{num}{
    number of trainig session. default value: 1
  }
}
\description{
  Get an object of class training and plot heart rate, speed and altitude over time of one training session.
}


