\docType{methods}
\name{plot_starts}
\alias{plot_starts}
\alias{plot_route, training-method}
\title{Plot of all start points of the training object}
\usage{
  trimp_exp(obj)

\S4method{plot_starts}{training}(obj)
}
\arguments{
  \item{obj}{
    an object of class training
  }
}
\description{
  Get an object of class training and plot all start points of the given training object. Click on the circle to geht the number of the session.
}

