\name{read_tcxToRun}
\alias{read_tcxToRun}
\title{
  creates a data frame from a given running activity saved in a tcx-file
}
\description{
  Get a tcx-file of a running activity and converts this in a data frame
}
\usage{
  read_tcxToRun(file)
}
\arguments{
  \item{file}{
    tcx file of a running activity
  }
}
\value{
  \item{df}{
    data frame of the running activity
  }
}
