\docType{methods}
\name{trimp_exp}
\alias{trimp_exp}
\alias{trimp_exp,training-method}
\title{calculate the exponantial Trimp of the activities}
\usage{
  trimp_exp(obj)

\S4method{trimp_exp}{training}(obj)
}
\arguments{
  \item{obj}{an object of class training}
}
\description{
  Get an object of class training and calculates the exponantial Trimp. Gives an output of the Trimp's over all activities.
}


