\docType{methods}
\name{trimp_exp_fct}
\alias{trimp_exp_fct}
\alias{trimp_exp_fct,training-method}
\title{
  Calculates the exponantial Trimp
}
\usage{
  trimp_exp_fct(obj)

\S4method{trimp_exp_fct}{training}(obj)
}
\arguments{
  \item{obj}{
    object of class training
  }
}
\value{
  \item{trimp_ex}{
    all exponantial Trimp's of the training activity
  }
}
\description{
  Get an object of class training and calculates to every training activity the exponantial Trimp
}
