\docType{methods}
\name{trimp_zone}
\alias{trimp_zone}
\alias{trimp_zone,training-method}
\title{calculate the zonal Trimp of the activities}
\usage{
  trimp_zone(train)

\S4method{trimp_zone}{training}(train)
}
\arguments{
  \item{train}{an object of class training}
}
\value{
  \item{trimpzo}{
    array with zonal Trimp of all activities
  }
}
\description{
  Get an object of class training and calculates the zonal Trimp. Gives additionally an output of the Trimp's over all     activities
}


