\docType{methods}
\name{plot_performance}
\alias{plot_performance}
\alias{plot_performance,training-method}
\title{Plot heart rate, speed and altitude over time}
\usage{
  plot_performance(obj, num = 1)

\S4method{plot_performance}{training}(obj, num = 1)
}
\arguments{
  \item{obj}{
    an object of class training
  }
  \item{num}{
    number of trainig session. default value: 1
  }
}
\description{
  Get an object of class training and plot heart rate, speed and altitude over time of one training session.
}


