\docType{methods}
\name{trimp_exp}
\alias{trimp_exp}
\alias{trimp_exp,training-method}
\title{calculate the exponantial Trimp of the activities}
\usage{
  trimp_exp(train)

\S4method{trimp_exp}{training}(train)
}
\arguments{
  \item{train}{an object of class training}
}
\description{
  Get an object of class training and calculates the exponantial Trimp. Gives additionally an output of the Trimp's over all     activities
}


