\docType{methods}
\name{trimp_zone}
\alias{trimp_zone}
\alias{trimp_zone,training-method}
\title{calculate the zonal Trimp of the activities}
\usage{
  trimp_zone(obj)

\S4method{trimp_zone}{training}(obj)
}
\arguments{
  \item{obj}{an object of class training}
}
\description{
  Get an object of class training and calculates the zonal Trimp. Gives an output of the Trimp's over all activities.
}


